+++ dronecan_gui_tool/panels/esc_panel.py
@@ -109,7 +109,7 @@ class ESCPanel(QDialog):
         self._msg_viewer.setHorizontalScrollBarPolicy(Qt.ScrollBarAsNeeded)
 
         self._bcast_timer = QTimer(self)
-        self._bcast_timer.start(self.DEFAULT_INTERVAL * 1e3)
+        self._bcast_timer.start(int(self.DEFAULT_INTERVAL * 1e3))
         self._bcast_timer.timeout.connect(self._do_broadcast)
 
         layout = QVBoxLayout(self)
